python/requirements.txt
web3~=6.8.0
discord-webhook~=1.2.1
absl-py~=1.4.0
python-dotenv~=1.0.0
